chat

+++ src/main/java/com/DAI/ProChild/Chat/ChatMessageRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatMessageRepository extends PagingAndSortingRepository<ChatMessage, Long> {
+public interface ChatMessageRepository extends PagingAndSortingRepository<ChatMessage, Integer> {
 }
